#5 WebSocket client modifications
Finished part of code responsible for receiving and displaying camera image (works)
Modified code responsible for sending steering commands (to be tested)

+++ AGVTesterApp/app/src/main/java/com/example/agvtesterapp/models/Header.kt
@@ -4,6 +4,6 @@ import java.io.Serializable
 
 data class Header(
     val seq: Int,
-    val stamp: Long,
+    val stamp: Double,
     val frameId: String
 ): Serializable
